\docType{data}
\name{dc_crime_data_sample}
\alias{dc_crime_data_sample}
\title{Sample DC Crime Data}
\description{
  Sample of 1000 crime records from the full Ward 1-8
  dataset. Sampled from data through 2012-07-17
}
\author{
  Erk Shilts \email{erik_shilts@gmail.com}
}
\keyword{data}
\keyword{datasets}

