\docType{data}
\name{dc_hybrid_zoom12}
\alias{dc_hybrid_zoom12}
\title{Hybrid (Road and Satellite) Map of DC at Zoom Level 12}
\description{
  Google road and satellite hybrid map of Washington DC at
  zoom level 12. Retrieved via get_dc_map() which uses
  ggmap::get_map().
}
\author{
  Erk Shilts \email{erik_shilts@gmail.com}
}
\keyword{data}
\keyword{datasets}

