\name{get_dc_color_names}
\alias{get_dc_color_names}
\title{Retrieves DC Crime Maps Color Names}
\usage{
  get_dc_color_names()
}
\value{
  Vector of names of DC Crime maps colors
}
\description{
  Returns a vector of the DC Crime Maps colors names for
  use in \{get_dc_hex\}
}
\examples{
get_dc_color_names()
}

