\name{get_dc_map}
\alias{get_dc_map}
\title{Get A Map of Washington DC}
\usage{
  get_dc_map(zoom = 12, maptype = "roadmap")
}
\arguments{
  \item{zoom}{zoom level (2-20)}

  \item{maptype}{type of map (e.g. roadmap, satellite,
  hybrid)}
}
\description{
  Returns a map of Washington, DC for creating plots with
  ggplot2 and ggmap. The user can set the zoom and map type
  for now.
}

