\name{plot_offense_counts}
\alias{plot_offense_counts}
\title{Plot Offense Counts of Crime in DC}
\usage{
  plot_offense_counts(crime_data)
}
\arguments{
  \item{crime_data}{data.frame to plot}
}
\description{
  Plots counts of crimes from 2006-2012.
}
\examples{
plot_offense_counts(crime_data=dc_crime_data_sample)
}

